Rebase, more customization and fixes
