+++ docs/components/page-higher-order/withHeadingManager.js
@@ -2,8 +2,10 @@ import * as React from 'react';
 
 export const HeadingsContext = React.createContext();
 
-export default Component => props => (
+const withHeadingManager = Component => props => (
   <HeadingsContext.Consumer>
     {headingManager => <Component headingManager={headingManager} {...props} />}
   </HeadingsContext.Consumer>
 );
+
+export default withHeadingManager;
